admin_service/requirements.txt
Mako==1.3.11
MarkupSafe==3.0.3
PyYAML==6.0.3
Pygments==2.20.0
PyJWT==2.12.1
SQLAlchemy==2.0.36
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2026.2.25
click==8.3.2
fastapi==0.115.6
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.1
pluggy==1.6.0
pydantic-settings==2.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.4.1
python-dotenv==1.2.2
starlette==0.41.3
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
